backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
sqlalchemy
psycopg2-binary